add connection-info package, move types and utils from connection-storage
