+++ packages/compass-connections/package.json
@@ -58,6 +58,7 @@
     "@mongodb-js/compass-connection-import-export": "^0.20.0",
     "@mongodb-js/compass-logging": "^1.2.9",
     "@mongodb-js/connection-form": "^1.21.0",
+    "@mongodb-js/connection-info": "^0.1.0",
     "compass-preferences-model": "^2.17.0",
     "hadron-ipc": "^3.2.7",
     "lodash": "^4.17.21",
