+++ packages/compass-connections/src/modules/telemetry.spec.ts
@@ -1,7 +1,7 @@
 import { once } from 'events';
 import { expect } from 'chai';
 import type { DataService } from 'mongodb-data-service';
-import type { ConnectionInfo } from '@mongodb-js/connection-storage/renderer';
+import type { ConnectionInfo } from '@mongodb-js/connection-info';
 import {
   trackConnectionAttemptEvent,
   trackNewConnectionEvent,
