+++ packages/compass-home/package.json
@@ -58,7 +58,7 @@
     "@mongodb-js/compass-sidebar": "^5.22.0",
     "@mongodb-js/compass-welcome": "^0.21.0",
     "@mongodb-js/compass-workspaces": "^0.4.0",
-    "@mongodb-js/connection-storage": "^0.7.5",
+    "@mongodb-js/connection-info": "^0.1.0",
     "compass-preferences-model": "^2.17.0",
     "hadron-app-registry": "^9.1.3",
     "hadron-ipc": "^3.2.7",
@@ -89,7 +89,7 @@
     "@mongodb-js/compass-sidebar": "^5.22.0",
     "@mongodb-js/compass-welcome": "^0.21.0",
     "@mongodb-js/compass-workspaces": "^0.4.0",
-    "@mongodb-js/connection-storage": "^0.7.5",
+    "@mongodb-js/connection-info": "^0.1.0",
     "compass-preferences-model": "^2.17.0",
     "hadron-app-registry": "^9.1.3",
     "hadron-ipc": "^3.2.7",
