+++ packages/compass-home/src/components/workspace.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { css } from '@mongodb-js/compass-components';
-import type { ConnectionInfo } from '@mongodb-js/connection-storage/renderer';
+import type { ConnectionInfo } from '@mongodb-js/connection-info';
 import { CompassShellPlugin } from '@mongodb-js/compass-shell';
 import {
   WorkspaceTab as CollectionWorkspace,
