+++ packages/compass-sidebar/package.json
@@ -64,6 +64,7 @@
     "@mongodb-js/compass-maybe-protect-connection-string": "^0.15.0",
     "@mongodb-js/compass-workspaces": "^0.4.0",
     "@mongodb-js/connection-form": "^1.21.0",
+    "@mongodb-js/connection-info": "^0.1.0",
     "@mongodb-js/connection-storage": "^0.7.5",
     "@mongodb-js/mongodb-redux-common": "^2.0.19",
     "compass-preferences-model": "^2.17.0",
@@ -80,6 +81,7 @@
     "@mongodb-js/compass-maybe-protect-connection-string": "^0.15.0",
     "@mongodb-js/compass-workspaces": "^0.4.0",
     "@mongodb-js/connection-form": "^1.21.0",
+    "@mongodb-js/connection-info": "^0.1.0",
     "@mongodb-js/connection-storage": "^0.7.5",
     "@mongodb-js/mongodb-redux-common": "^2.0.19",
     "compass-preferences-model": "^2.17.0",
