+++ packages/compass-sidebar/src/components/sidebar.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback, useState } from 'react';
 import { cloneDeep } from 'lodash';
 import { connect } from 'react-redux';
-import { getConnectionTitle } from '@mongodb-js/connection-storage/renderer';
-import type { ConnectionInfo } from '@mongodb-js/connection-storage/renderer';
+import { getConnectionTitle } from '@mongodb-js/connection-info';
+import type { ConnectionInfo } from '@mongodb-js/connection-info';
 import {
   css,
   spacing,
