+++ packages/compass/src/app/auto-connect.ts
@@ -1,5 +1,5 @@
 import { ConnectionStorage } from '@mongodb-js/connection-storage/renderer';
-import type { ConnectionInfo } from '@mongodb-js/connection-storage/renderer';
+import type { ConnectionInfo } from '@mongodb-js/connection-info';
 import { promises as fsPromises } from 'fs';
 import { UUID } from 'bson';
 import { ipcRenderer } from 'hadron-ipc';
