+++ packages/connection-form/package.json
@@ -54,6 +54,7 @@
   "dependencies": {
     "@mongodb-js/compass-components": "^1.21.0",
     "@mongodb-js/compass-editor": "^0.20.0",
+    "@mongodb-js/connection-info": "^0.1.0",
     "@testing-library/react-hooks": "^7.0.2",
     "lodash": "^4.17.21",
     "mongodb-build-info": "^1.7.0",
@@ -61,7 +62,6 @@
     "mongodb-query-parser": "^4.0.0"
   },
   "devDependencies": {
-    "@mongodb-js/connection-storage": "^0.7.5",
     "@mongodb-js/eslint-config-compass": "^1.0.13",
     "@mongodb-js/mocha-config-compass": "^1.3.4",
     "@mongodb-js/prettier-config-compass": "^1.0.1",
