+++ packages/connection-form/src/utils/validation.spec.ts
@@ -4,7 +4,7 @@ import {
   validateConnectionOptionsErrors,
   validateConnectionOptionsWarnings,
 } from './validation';
-import type { ConnectionInfo } from '@mongodb-js/connection-storage/renderer';
+import type { ConnectionInfo } from '@mongodb-js/connection-info';
 
 describe('validation', function () {
   describe('Form Validation Errors', function () {
