+++ packages/connection-info/.prettierignore
@@ -1,0 +1,3 @@
+.nyc_output
+dist
+coverage
