+++ packages/connection-info/.prettierrc.json
@@ -1,0 +1,1 @@
+"@mongodb-js/prettier-config-compass"
