+++ packages/connection-info/src/connection-info.ts
