+++ packages/connection-info/src/connection-secrets.spec.ts
