+++ packages/connection-info/src/connection-secrets.ts
