+++ packages/connection-info/src/index.ts
@@ -1,0 +1,7 @@
+export type { ConnectionSecrets } from './connection-secrets';
+export { extractSecrets, mergeSecrets } from './connection-secrets';
+export type {
+  ConnectionInfo,
+  ConnectionFavoriteOptions,
+} from './connection-info';
+export { getConnectionTitle } from './connection-title';
