+++ packages/connection-storage/src/main.ts
@@ -1,9 +1,5 @@
 export { ConnectionStorage } from './connection-storage';
-export type { ConnectionSecrets } from './connection-secrets';
-export { extractSecrets, mergeSecrets } from './connection-secrets';
 export type {
   ExportConnectionOptions,
   ImportConnectionOptions,
 } from './import-export-connection';
-export type { ConnectionInfo } from './connection-info';
-export { getConnectionTitle } from './connection-title';
