+++ packages/connection-storage/src/renderer.ts
@@ -1,11 +1,6 @@
 import { ipcRenderer } from 'hadron-ipc';
 
 import type { ConnectionStorage as ConnectionStorageMain } from './connection-storage';
-export type {
-  ConnectionInfo,
-  ConnectionFavoriteOptions,
-} from './connection-info';
-export { getConnectionTitle } from './connection-title';
 
 export class ConnectionStorage {
   private static _ipc = ipcRenderer?.createInvoke<
